display: Create common chromatic adaptation tint controller
Display white balance is not the only use case for chromatic adaptation
and most of the code can be reused, so create a common tint controller
class for chromatic adaptation.
